Move healthcare to testAnalyticsClient.spec.ts
